#1: split operators into modules
