+++ doc/html/genindex.html
@@ -85,13 +85,11 @@
   <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="index.html#module-ffmpeg">ffmpeg (module)</a>
 </li>
-      <li><a href="index.html#ffmpeg.file_input">file_input() (in module ffmpeg)</a>
-</li>
   </ul></td>
   <td style="width: 33%; vertical-align: top;"><ul>
-      <li><a href="index.html#ffmpeg.file_output">file_output() (in module ffmpeg)</a>
+      <li><a href="index.html#ffmpeg.file_input">file_input() (in module ffmpeg)</a>
 </li>
-      <li><a href="index.html#ffmpeg.fluent">fluent() (in module ffmpeg)</a>
+      <li><a href="index.html#ffmpeg.file_output">file_output() (in module ffmpeg)</a>
 </li>
   </ul></td>
 </tr></table>
