+++ doc/html/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["index"],envversion:52,filenames:["index.rst"],objects:{"":{ffmpeg:[0,0,0,"-"]},ffmpeg:{colorchannelmixer:[0,1,1,""],concat:[0,1,1,""],drawbox:[0,1,1,""],file_input:[0,1,1,""],file_output:[0,1,1,""],fluent:[0,1,1,""],get_args:[0,1,1,""],hflip:[0,1,1,""],hue:[0,1,1,""],merge_outputs:[0,1,1,""],overlay:[0,1,1,""],overwrite_output:[0,1,1,""],run:[0,1,1,""],setpts:[0,1,1,""],trim:[0,1,1,""],vflip:[0,1,1,""],zoompan:[0,1,1,""]}},objnames:{"0":["py","module","Python module"],"1":["py","function","Python function"]},objtypes:{"0":"py:module","1":"py:function"},terms:{"class":0,"function":0,arg:0,bind:0,cmd:0,color:0,colorchannelmix:0,concat:0,correspond:0,decor:0,drawbox:0,eof_act:0,expr:0,file:0,file_input:0,file_output:0,filenam:0,fixm:0,fluent:0,from:0,get_arg:0,height:0,hflip:0,hue:0,index:0,input:0,instanc:0,kwarg:0,list:0,main_par:0,make:0,merge_output:0,method:0,modul:0,node:0,node_class:0,none:0,option:0,overlai:0,overlay_par:0,overwrite_output:0,page:0,paramet:0,parent:0,repeat:0,run:0,search:0,set:0,setpt:0,show:0,specifi:0,thick:0,trim:0,vflip:0,width:0,zoompan:0},titles:["Welcome to ffmpeg-python\u2019s documentation!"],titleterms:{document:0,ffmpeg:0,indic:0,python:0,tabl:0,welcom:0}})
+Search.setIndex({docnames:["index"],envversion:52,filenames:["index.rst"],objects:{"":{ffmpeg:[0,0,0,"-"]},ffmpeg:{colorchannelmixer:[0,1,1,""],concat:[0,1,1,""],drawbox:[0,1,1,""],file_input:[0,1,1,""],file_output:[0,1,1,""],get_args:[0,1,1,""],hflip:[0,1,1,""],hue:[0,1,1,""],merge_outputs:[0,1,1,""],overlay:[0,1,1,""],overwrite_output:[0,1,1,""],run:[0,1,1,""],setpts:[0,1,1,""],trim:[0,1,1,""],vflip:[0,1,1,""],zoompan:[0,1,1,""]}},objnames:{"0":["py","module","Python module"],"1":["py","function","Python function"]},objtypes:{"0":"py:module","1":"py:function"},terms:{"class":[],"function":[],adjust:0,arg:0,base:[],bind:[],channel:0,cmd:0,color:0,colorchannelmix:0,concat:0,correspond:[],decor:[],drawbox:0,eof_act:0,expr:0,file:[],file_input:0,file_output:0,filenam:0,filter:0,filternod:[],fixm:[],fluent:[],frame:0,from:[],get_arg:0,globalnod:[],height:0,hflip:0,hue:0,index:0,input:0,inputnod:[],instanc:[],kwarg:0,list:[],main_par:0,make:[],merge_output:0,method:[],mix:0,modul:0,name:[],node:[],node_class:[],none:0,object:[],oper:[],option:[],outputnod:[],overlai:0,overlay_par:0,overwrite_output:0,page:0,paramet:[],parent:0,repeat:0,run:0,search:0,set:[],setpt:0,show:[],specifi:[],thick:0,trim:0,vflip:0,video:0,width:0,zoompan:0},titles:["ffmpeg-python: Python bindings for FFmpeg"],titleterms:{bind:0,document:[],ffmpeg:0,indic:0,python:0,tabl:0,welcom:[]}})
