+++ ffmpeg/tests/test_ffmpeg.py
@@ -1,4 +1,3 @@
-
 import ffmpeg
 import os
 import subprocess
